This is intermediate code on the path to version 2.0. It allows multiple charts and creation of Links. Links allow unions between people on different charts and prevent long crossing lines on a single chart.
This code allows use of User-Defined Properties to record and chart up to 3 kinds of adoptions (user-defined). Adoption links between kin are normal, first-class ties, although of lower priority than blood ties.

I have released a limited edition of this code as version 1.2. It is available in the downloads area.